requirements.txt
boto3==1.9.207
botocore==1.12.207
celery==4.4.0
django==2.2.17
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-ckeditor==5.7.1
django-configurations==2.1
django-extensions==2.2.6
django-jet==1.0.8
django-jsonfield==1.4.0
django-mysql==3.3.0
django-relative-softdeletion==0.0.2
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
django-storages==1.7.1
drf-yasg==1.20.0
furl==2.0.0
importlib-metadata==1.4.0
kombu==4.6.7
mysqlclient==1.4.6
Pillow>=7.1.0
PyJWT==1.7.1
python-crontab==2.3.8
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.3
redis==3.3.11
requests==2.24.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.24.3
vine==1.3.0
xlwt==1.3.0
zipp==2.1.0
